requirements.txt
apturl==0.5.2
asgiref==3.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.0
django-extensions==3.2.0
django-filter==22.1
django-guardian==2.4.0
djangorestframework==3.13.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
gpg===1.16.0-unknown
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.4.1
monotonic==1.6
more-itertools==8.10.0
natsort==8.0.2
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.12.0
PyGObject==3.42.0
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.8
requests==2.28.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
systemd-python==234
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
PyYAML == 6.0
uritemplate == 4.1.1
Markdown == 3.4.1
Pygments == 2.12.0
django-filter == 22.1
django-guardian == 2.4.0
djangorestframework == 3.13.1
django-extensions==3.2.0
django==4.0.0